% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_missing}
\alias{install_missing}
\title{Install missing packages}
\usage{
install_missing(pkg_list = "packages.yml")
}
\arguments{
\item{pkg_list}{path to the `packages.yml` file (defaults to `packages.yml`)}
}
\value{
a tibble with an updated `is_installed` column to confirm the success of all installations.
}
\description{
Once the required packages (defined in `packages.yml`) are listed and checked, installs the missing ones.
}
