% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{pkg_list}
\alias{pkg_list}
\title{Lists the required packages}
\usage{
pkg_list(path = ".", install = FALSE, scan = TRUE, rmd_dir = c("TD",
  "lectures", "site"))
}
\arguments{
\item{path}{character vector pointing to the root of the bs2 repository containing the optional `packages.yml` file.}

\item{install}{boolean telling whether to install missing packages (defaults to `FALSE`)}

\item{scan}{boolean whether to scan source files for used packages (defaults to `TRUE`)}

\item{rmd_dir}{character vector containing the subfolders with the Rmd source files.}
}
\value{
A tibble listing the package, the version, whether it is already installed or not.
}
\description{
The `packages.yml` file contains a list of packages that are required.
}
