% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{pkg_scan}
\alias{pkg_scan}
\title{Scan the site files for used packages}
\usage{
pkg_scan(path, recursive = TRUE)
}
\arguments{
\item{path}{folders containing the Rmd files.}

\item{recursive}{logical should Rmd files be searched recursively}
}
\value{
path to the Rmd files
}
\description{
Will extract all loaded packages using `library()`, `require()` or direct calls `::`.
}
