% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{scan_packages}
\alias{scan_packages}
\title{Scan the site files for used packages}
\usage{
scan_packages(path = c("TD", "lectures", "site"))
}
\arguments{
\item{path}{folders containing the Rmd files (defaults to `c("TD", "lectures", "site")`).}
}
\value{
path to the Rmd files
}
\description{
Will extract all loaded packages using `library()`, `require()` or direct calls `::`.
}
