% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_colours.R
\name{use_box_colours}
\alias{use_box_colours}
\title{Use box colours defined for the rworkshop or biostat2 site}
\usage{
use_box_colours()
}
\description{
Generates the `css/box.css` file and changes the css setting in `lectures/_output.yml` if required.
}
