% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{wanted_pkg}
\alias{wanted_pkg}
\title{Lists the required packages}
\usage{
wanted_pkg(pkg_list = "packages.yml")
}
\arguments{
\item{pkg_list}{path to the `packages.yml` file (defaults to `packages.yml`)}
}
\value{
A tibble listing the package, the version, whether it is already installed or not.
}
\description{
The `packages.yml` file contains a list of packages that are required.
}
