% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_rds.R
\name{write_cache}
\alias{write_cache}
\title{Write objects to the website template cache folder}
\usage{
write_cache(object, cache_folder = "_cache", ...)
}
\arguments{
\item{object}{object to save.}

\item{cache_folder}{defaults to `_cache` subfolder.}

\item{...}{Additional parameters passed to `readr::write_rds()` such as compression arguments.}
}
\value{
`write_cache()` returns the input object invisibly
}
\description{
Use this function to store objects as rds files in a standardised location and speed up knitting.
}
