% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{create_navbar}
\alias{create_navbar}
\title{Create a navbar with the branding logos}
\usage{
create_navbar(site_yml = file.path("site", "site.yml"), logo_img)
}
\arguments{
\item{site_yml}{path to the `_site.yml` file.}

\item{logo_img}{path to the logo files}
}
\description{
Reads the _site.yml to render the _navbar.html and replaces the navbar title by the branding logos.
}
