% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_missing}
\alias{install_missing}
\title{Install missing packages}
\usage{
install_missing(.df, do_install = isTRUE(Sys.getenv("BS2_DEPLOY") == "1"))
}
\arguments{
\item{.df}{A tibble generated by `wanted_pkg()`}

\item{do_install}{a boolean enabling the installation (defaults to `isTRUE(Sys.getenv("BS2_DEPLOY") == "1")`)}
}
\value{
The `.df` tibble with an updated `is_installed` column to confirm the success of all installations.
}
\description{
Once the required packages (defined in `packages.yml`) are listed and checked, installs the missing ones.
}
